requirements.txt
constriction
numpy
pandas
pillow
pytorch-msssim
PyYAML
scikit-learn
scipy
tqdm
vector-quantize-pytorch